CC-88 Refactor and implement more API features
Lots of changes in this commit. I should've done these as separate
commits but the code has been highly volatile as I've been exploring
with different ways to implement this.
1) Code now uses three types of objects.
- The `GGB` object carries global defaults and has methods for creating
abstract applets as objects called `Blueprint`s.
- The `Blueprint` object has a `draw` method for drawing an applet onto
the page as an object called a `Drawing`.
- The `Drawing` object has methods for making API calls to that drawing.
This solves the problem of being able to draw multiple copies of the
same applet to the page, allows one to set options at either the global
defaults level or blueprint level and to see which options a drawing
was drawn with. Moreover having an object for each drawing makes it
possible to make different API calls to different copies of the same
applet.

2) Function calls no longer need to be terminated with semicolons.

3) The `Drawing` object has `set` and `get` methods for making calls
to Geogebra's API. The `get` methods take a python variable name as a
string and after they're called the result of the API call will be
stored in said Python variable. These are mostly a proof of concept. I
plan to use some metaprogramming techniques to implement each of these
as its own Python method with its own help info. The `get` methods work
by using Javascript to tell the Jupyter kernel to execute Python code
behind the scene. This feature can be extended but has some limitations.
Strings are not currently escaped when moving between JS and Python and
vice versa, so this may cause breakage in some cases.

4) Option handling has been abstracted to an options object. Should make
it easier to implement better debug and pretty printing functionality.

5) The notebook has had some minor changes reflecting the minor changes
to the library. The big one being that `draw()` no longer returns the
original object. A bigger rewrite is forthcoming.