CC-88 Hotfix: Set html5NoWebSimple to false
Enable `web Simple` in Geogebra's deploygg.js as a workaround to address
issues where 'On Click' interaction fails to work for some worksheets.

There is very little documentation on `web Simple` but it seems that
it is not meant to be used in all cases and deployggb.js will try to
disable it in some cases depending on the applet and options loaded.