CC-209 Improve support for Run All Cells
ggb.py will now keep track of whether or not the javascript library
has finished downloading and will wait to load cells for a finite
amount of time if the javascript is not ready. There is a bug when
running cells containing ggb.app() applets, it's likely a race
condition and more investigation is necessary. An error message will
now be displayed if the javascript library fails to download, for
instance if there is no internet connection. More messages involving
javascript loading will be shown in the javascript console.