git-nbrstrip: rebuild commits on the first ancestor of our git range
Previously, git-nbrstrip would rebuild new versions of commits on top
of @{upstream}, but if @{upstream} has moved forward from the time we
started this work-in-progress, git-nbrstrip would effectively perform
a rebase.

Since that's beyond the scope of what the script is trying to do, we
explicitly make sure we're rebuilding commits on top of where they
started.