git-nbrstrip: add all changes in a commit, fixing issue #56
Previously, git-nbrstrip would not catch changes in files that didn't
exist before the range we were working within (or for that matter,
files which existed, dissapeared, and came back within that range).

Doing a blanket 'git add .' resolves this issue. It does mean that we
need to be more strict about not allowing the script to operate when
there's any unclean files in the repository, even if they're
untracked.