git-nbrstrip: continue but warn if a recreated commit doesn't change anything
I've seen this happen in cases where a output-stripping commit is in
the history, and we're trying to clean it up. By its nature, it's not
doing anything more than what git-nbrstrip is doing, therefore nothing
happens.

That was actually causing the script to fail -- but the faulty
error-catching (corrected in the previous commit) was masking it.